requirements.txt
facenet-pytorch
opencv-python
gTTS
playsound
SpeechRecognition

datasets
transformers
torch 
torchvision 
torchaudio
PyAudio
#PyGObject
accelerate 
peft 
bitsandbytes 
transformers
scipy
trl

google-api-python-client
google-auth-httplib2
google-auth-oauthlib